tests: Breakup layer_validation_tests into more files
