+++ tests/CMakeLists.txt
@@ -41,6 +41,12 @@ target_sources(vk_layer_validation_tests PRIVATE
     framework/ray_tracing_nv.cpp
     framework/external_memory_sync.h
     framework/external_memory_sync.cpp
+    framework/barrier_queue_family.h
+    framework/barrier_queue_family.cpp
+    framework/descriptor_helper.h
+    framework/descriptor_helper.cpp
+    framework/thread_timeout_helper.h
+    framework/thread_timeout_helper.cpp
     unit/amd_best_practices.cpp
     unit/android_hardware_buffer.cpp
     unit/android_hardware_buffer_positive.cpp
