+++ tests/framework/layer_validation_tests.cpp
@@ -183,29 +183,6 @@ void UpdateDescriptor(ThreadTestData *data) {
 
 #endif  // GTEST_IS_THREADSAFE
 
-bool ThreadTimeoutHelper::WaitForThreads(int timeout_in_seconds) {
-    std::unique_lock lock(mutex_);
-    return cv_.wait_for(lock, std::chrono::seconds{timeout_in_seconds}, [this] {
-        std::lock_guard lock_guard(active_thread_mutex_);
-        return active_threads_ == 0;
-    });
-}
-
-void ThreadTimeoutHelper::OnThreadDone() {
-    bool last_worker = false;
-    {
-        std::lock_guard lock(active_thread_mutex_);
-        active_threads_--;
-        assert(active_threads_ >= 0);
-        if (!active_threads_) {
-            last_worker = true;
-        }
-    }
-    if (last_worker) {
-        cv_.notify_one();
-    }
-}
-
 void ReleaseNullFence(ThreadTestData *data) {
     for (int i = 0; i < 40000; i++) {
         vk::DestroyFence(data->device, VK_NULL_HANDLE, NULL);
@@ -771,388 +748,6 @@ void SetImageLayout(vkt::Device *device, VkImageAspectFlags aspect, VkImage imag
     cmd_buf.end();
 
     cmd_buf.QueueCommandBuffer();
-}
-
-std::unique_ptr<VkImageObj> VkArmBestPracticesLayerTest::CreateImage(VkFormat format, const uint32_t width, const uint32_t height,
-                                                                     VkImageUsageFlags attachment_usage) {
-    auto img = std::unique_ptr<VkImageObj>(new VkImageObj(m_device));
-    img->Init(width, height, 1, format,
-              VK_IMAGE_USAGE_SAMPLED_BIT | attachment_usage | VK_IMAGE_USAGE_TRANSFER_SRC_BIT | VK_IMAGE_USAGE_TRANSFER_DST_BIT,
-              VK_IMAGE_TILING_OPTIMAL);
-    return img;
-}
-
-VkRenderPass VkArmBestPracticesLayerTest::CreateRenderPass(VkFormat format, VkAttachmentLoadOp load_op,
-                                                           VkAttachmentStoreOp store_op) {
-    VkRenderPass renderpass{VK_NULL_HANDLE};
-
-    // Create renderpass
-    VkAttachmentDescription attachment = {};
-    attachment.format = format;
-    attachment.samples = VK_SAMPLE_COUNT_1_BIT;
-    attachment.loadOp = load_op;
-    attachment.storeOp = store_op;
-    attachment.initialLayout = VK_IMAGE_LAYOUT_UNDEFINED;
-    attachment.finalLayout = VK_IMAGE_LAYOUT_GENERAL;
-
-    VkAttachmentReference attachment_reference = {};
-    attachment_reference.attachment = 0;
-    attachment_reference.layout = VK_IMAGE_LAYOUT_GENERAL;
-
-    VkSubpassDescription subpass = {};
-    subpass.colorAttachmentCount = 1;
-    subpass.pColorAttachments = &attachment_reference;
-
-    VkRenderPassCreateInfo rpinf = vku::InitStructHelper();
-    rpinf.attachmentCount = 1;
-    rpinf.pAttachments = &attachment;
-    rpinf.subpassCount = 1;
-    rpinf.pSubpasses = &subpass;
-    rpinf.dependencyCount = 0;
-    rpinf.pDependencies = nullptr;
-
-    VkResult result = vk::CreateRenderPass(m_device->handle(), &rpinf, nullptr, &renderpass);
-    assert(result == VK_SUCCESS);
-    (void)result;
-
-    return renderpass;
-}
-
-VkFramebuffer VkArmBestPracticesLayerTest::CreateFramebuffer(const uint32_t width, const uint32_t height, VkImageView image_view,
-                                                             VkRenderPass renderpass) {
-    VkFramebuffer framebuffer{VK_NULL_HANDLE};
-
-    VkFramebufferCreateInfo framebuffer_create_info = vku::InitStructHelper();
-    framebuffer_create_info.renderPass = renderpass;
-    framebuffer_create_info.attachmentCount = 1;
-    framebuffer_create_info.pAttachments = &image_view;
-    framebuffer_create_info.width = width;
-    framebuffer_create_info.height = height;
-    framebuffer_create_info.layers = 1;
-
-    VkResult result = vk::CreateFramebuffer(m_device->handle(), &framebuffer_create_info, nullptr, &framebuffer);
-    assert(result == VK_SUCCESS);
-    (void)result;
-
-    return framebuffer;
-}
-
-VkSampler VkArmBestPracticesLayerTest::CreateDefaultSampler() {
-    VkSampler sampler{VK_NULL_HANDLE};
-
-    VkSamplerCreateInfo sampler_create_info = vku::InitStructHelper();
-    sampler_create_info.magFilter = VK_FILTER_NEAREST;
-    sampler_create_info.minFilter = VK_FILTER_NEAREST;
-    sampler_create_info.mipmapMode = VK_SAMPLER_MIPMAP_MODE_NEAREST;
-    sampler_create_info.addressModeU = VK_SAMPLER_ADDRESS_MODE_REPEAT;
-    sampler_create_info.addressModeV = VK_SAMPLER_ADDRESS_MODE_REPEAT;
-    sampler_create_info.addressModeW = VK_SAMPLER_ADDRESS_MODE_REPEAT;
-    sampler_create_info.borderColor = VK_BORDER_COLOR_FLOAT_OPAQUE_WHITE;
-    sampler_create_info.maxLod = VK_LOD_CLAMP_NONE;
-
-    VkResult result = vk::CreateSampler(m_device->handle(), &sampler_create_info, nullptr, &sampler);
-    assert(result == VK_SUCCESS);
-    (void)result;
-
-    return sampler;
-}
-
-OneOffDescriptorSet::OneOffDescriptorSet(vkt::Device *device, const Bindings &bindings,
-                                         VkDescriptorSetLayoutCreateFlags layout_flags, void *layout_pnext,
-                                         VkDescriptorPoolCreateFlags poolFlags, void *allocate_pnext)
-    : device_{device}, pool_{}, layout_(*device, bindings, layout_flags, layout_pnext), set_(VK_NULL_HANDLE) {
-    VkResult err;
-    std::vector<VkDescriptorPoolSize> sizes;
-    for (const auto &b : bindings) sizes.push_back({b.descriptorType, std::max(1u, b.descriptorCount)});
-
-    VkDescriptorPoolCreateInfo dspci = {
-        VK_STRUCTURE_TYPE_DESCRIPTOR_POOL_CREATE_INFO, nullptr, poolFlags, 1, uint32_t(sizes.size()), sizes.data()};
-    err = vk::CreateDescriptorPool(device_->handle(), &dspci, nullptr, &pool_);
-    if (err != VK_SUCCESS) return;
-
-    if ((layout_flags & VK_DESCRIPTOR_SET_LAYOUT_CREATE_PUSH_DESCRIPTOR_BIT_KHR) == 0) {
-        VkDescriptorSetAllocateInfo alloc_info = {VK_STRUCTURE_TYPE_DESCRIPTOR_SET_ALLOCATE_INFO, allocate_pnext, pool_, 1,
-                                                  &layout_.handle()};
-        err = vk::AllocateDescriptorSets(device_->handle(), &alloc_info, &set_);
-    }
-}
-
-OneOffDescriptorSet::~OneOffDescriptorSet() {
-    // No need to destroy set-- it's going away with the pool.
-    vk::DestroyDescriptorPool(device_->handle(), pool_, nullptr);
-}
-
-bool OneOffDescriptorSet::Initialized() { return pool_ != VK_NULL_HANDLE && layout_.initialized() && set_ != VK_NULL_HANDLE; }
-
-void OneOffDescriptorSet::Clear() {
-    resource_infos.clear();
-    descriptor_writes.clear();
-}
-
-void OneOffDescriptorSet::AddDescriptorWrite(uint32_t binding, uint32_t array_element, VkDescriptorType descriptor_type) {
-    VkWriteDescriptorSet descriptor_write = vku::InitStructHelper();
-    descriptor_write.dstSet = set_;
-    descriptor_write.dstBinding = binding;
-    descriptor_write.dstArrayElement = array_element;
-    descriptor_write.descriptorCount = 1;
-    descriptor_write.descriptorType = descriptor_type;
-    descriptor_writes.emplace_back(descriptor_write);
-}
-
-void OneOffDescriptorSet::WriteDescriptorBufferInfo(int binding, VkBuffer buffer, VkDeviceSize offset, VkDeviceSize range,
-                                                    VkDescriptorType descriptorType, uint32_t arrayElement) {
-    VkDescriptorBufferInfo buffer_info = {};
-    buffer_info.buffer = buffer;
-    buffer_info.offset = offset;
-    buffer_info.range = range;
-
-    ResourceInfo resource_info;
-    resource_info.buffer_info = buffer_info;
-    resource_infos.emplace_back(resource_info);
-    AddDescriptorWrite(binding, arrayElement, descriptorType);
-}
-
-void OneOffDescriptorSet::WriteDescriptorBufferView(int binding, VkBufferView buffer_view, VkDescriptorType descriptorType,
-                                                    uint32_t arrayElement) {
-    ResourceInfo resource_info;
-    resource_info.buffer_view = buffer_view;
-    resource_infos.emplace_back(resource_info);
-    AddDescriptorWrite(binding, arrayElement, descriptorType);
-}
-
-void OneOffDescriptorSet::WriteDescriptorImageInfo(int binding, VkImageView image_view, VkSampler sampler,
-                                                   VkDescriptorType descriptorType, VkImageLayout imageLayout,
-                                                   uint32_t arrayElement) {
-    VkDescriptorImageInfo image_info = {};
-    image_info.imageView = image_view;
-    image_info.sampler = sampler;
-    image_info.imageLayout = imageLayout;
-
-    ResourceInfo resource_info;
-    resource_info.image_info = image_info;
-    resource_infos.emplace_back(resource_info);
-    AddDescriptorWrite(binding, arrayElement, descriptorType);
-}
-
-void OneOffDescriptorSet::UpdateDescriptorSets() {
-    assert(resource_infos.size() == descriptor_writes.size());
-    for (size_t i = 0; i < resource_infos.size(); i++) {
-        const auto &info = resource_infos[i];
-        if (info.image_info.has_value()) {
-            descriptor_writes[i].pImageInfo = &info.image_info.value();
-        } else if (info.buffer_info.has_value()) {
-            descriptor_writes[i].pBufferInfo = &info.buffer_info.value();
-        } else {
-            assert(info.buffer_view.has_value());
-            descriptor_writes[i].pTexelBufferView = &info.buffer_view.value();
-        }
-    }
-    vk::UpdateDescriptorSets(device_->handle(), descriptor_writes.size(), descriptor_writes.data(), 0, NULL);
-}
-
-BarrierQueueFamilyBase::QueueFamilyObjs::~QueueFamilyObjs() {
-    delete command_buffer2;
-    delete command_buffer;
-    delete command_pool;
-    delete queue;
-}
-
-void BarrierQueueFamilyBase::QueueFamilyObjs::Init(vkt::Device *device, uint32_t qf_index, VkQueue qf_queue,
-                                                   VkCommandPoolCreateFlags cp_flags) {
-    index = qf_index;
-    queue = new vkt::Queue(qf_queue, qf_index);
-    command_pool = new vkt::CommandPool(*device, qf_index, cp_flags);
-    command_buffer = new vkt::CommandBuffer(device, command_pool, VK_COMMAND_BUFFER_LEVEL_PRIMARY, queue);
-    command_buffer2 = new vkt::CommandBuffer(device, command_pool, VK_COMMAND_BUFFER_LEVEL_PRIMARY, queue);
-}
-
-BarrierQueueFamilyBase::Context::Context(VkLayerTest *test, const std::vector<uint32_t> &queue_family_indices) : layer_test(test) {
-    if (0 == queue_family_indices.size()) {
-        return;  // This is invalid
-    }
-    vkt::Device *device_obj = layer_test->DeviceObj();
-    queue_families.reserve(queue_family_indices.size());
-    default_index = queue_family_indices[0];
-    for (auto qfi : queue_family_indices) {
-        VkQueue queue = device_obj->queue_family_queues(qfi)[0]->handle();
-        queue_families.emplace(std::make_pair(qfi, QueueFamilyObjs()));
-        queue_families[qfi].Init(device_obj, qfi, queue, VK_COMMAND_POOL_CREATE_RESET_COMMAND_BUFFER_BIT);
-    }
-    Reset();
-}
-
-void BarrierQueueFamilyBase::Context::Reset() {
-    layer_test->DeviceObj()->wait();
-    for (auto &qf : queue_families) {
-        vk::ResetCommandPool(layer_test->device(), qf.second.command_pool->handle(), 0);
-    }
-}
-
-void BarrierQueueFamilyTestHelper::Init(std::vector<uint32_t> *families, bool image_memory, bool buffer_memory) {
-    vkt::Device *device_obj = context_->layer_test->DeviceObj();
-
-    image_.Init(32, 32, 1, VK_FORMAT_B8G8R8A8_UNORM, VK_IMAGE_USAGE_COLOR_ATTACHMENT_BIT, VK_IMAGE_TILING_OPTIMAL, 0, families,
-                image_memory);
-
-    ASSERT_TRUE(image_.initialized());
-
-    image_barrier_ = image_.image_memory_barrier(VK_ACCESS_TRANSFER_READ_BIT, VK_ACCESS_TRANSFER_READ_BIT, image_.Layout(),
-                                                 image_.Layout(), image_.subresource_range(VK_IMAGE_ASPECT_COLOR_BIT, 0, 1, 0, 1));
-
-    VkMemoryPropertyFlags mem_prop = VK_MEMORY_PROPERTY_HOST_COHERENT_BIT;
-    auto buffer_ci = vkt::Buffer::create_info(256, VK_BUFFER_USAGE_TRANSFER_SRC_BIT | VK_BUFFER_USAGE_TRANSFER_DST_BIT, families);
-    if (buffer_memory) {
-        buffer_.init(*device_obj, buffer_ci, mem_prop);
-    } else {
-        buffer_.init_no_mem(*device_obj, buffer_ci);
-    }
-    ASSERT_TRUE(buffer_.initialized());
-    buffer_barrier_ = buffer_.buffer_memory_barrier(VK_ACCESS_TRANSFER_READ_BIT, VK_ACCESS_TRANSFER_READ_BIT, 0, VK_WHOLE_SIZE);
-}
-
-void Barrier2QueueFamilyTestHelper::Init(std::vector<uint32_t> *families, bool image_memory, bool buffer_memory) {
-    vkt::Device *device_obj = context_->layer_test->DeviceObj();
-
-    image_.Init(32, 32, 1, VK_FORMAT_B8G8R8A8_UNORM, VK_IMAGE_USAGE_COLOR_ATTACHMENT_BIT, VK_IMAGE_TILING_OPTIMAL, 0, families,
-                image_memory);
-
-    ASSERT_TRUE(image_.initialized());
-
-    image_barrier_ = image_.image_memory_barrier(VK_PIPELINE_STAGE_TRANSFER_BIT, VK_PIPELINE_STAGE_TRANSFER_BIT,
-                                                 VK_ACCESS_TRANSFER_READ_BIT, VK_ACCESS_TRANSFER_READ_BIT, image_.Layout(),
-                                                 image_.Layout(), image_.subresource_range(VK_IMAGE_ASPECT_COLOR_BIT, 0, 1, 0, 1));
-
-    VkMemoryPropertyFlags mem_prop = VK_MEMORY_PROPERTY_HOST_COHERENT_BIT;
-    auto buffer_ci = vkt::Buffer::create_info(256, VK_BUFFER_USAGE_TRANSFER_SRC_BIT | VK_BUFFER_USAGE_TRANSFER_DST_BIT, families);
-    if (buffer_memory) {
-        buffer_.init(*device_obj, buffer_ci, mem_prop);
-    } else {
-        buffer_.init_no_mem(*device_obj, buffer_ci);
-    }
-    ASSERT_TRUE(buffer_.initialized());
-    buffer_barrier_ = buffer_.buffer_memory_barrier(VK_PIPELINE_STAGE_TRANSFER_BIT, VK_PIPELINE_STAGE_TRANSFER_BIT,
-                                                    VK_ACCESS_TRANSFER_READ_BIT, VK_ACCESS_TRANSFER_READ_BIT, 0, VK_WHOLE_SIZE);
-}
-
-BarrierQueueFamilyBase::QueueFamilyObjs *BarrierQueueFamilyBase::GetQueueFamilyInfo(Context *context, uint32_t qfi) {
-    QueueFamilyObjs *qf;
-
-    auto qf_it = context->queue_families.find(qfi);
-    if (qf_it != context->queue_families.end()) {
-        qf = &(qf_it->second);
-    } else {
-        qf = &(context->queue_families[context->default_index]);
-    }
-    return qf;
-}
-
-void BarrierQueueFamilyTestHelper::operator()(const std::string &img_err, const std::string &buf_err, uint32_t src, uint32_t dst,
-                                              uint32_t queue_family_index, Modifier mod) {
-    auto &monitor = context_->layer_test->Monitor();
-    const bool has_img_err = img_err.size() > 0;
-    const bool has_buf_err = buf_err.size() > 0;
-    bool positive = !has_img_err && !has_buf_err;
-    if (has_img_err) monitor.SetDesiredFailureMsg(kErrorBit | kWarningBit, img_err);
-    if (has_buf_err) monitor.SetDesiredFailureMsg(kErrorBit | kWarningBit, buf_err);
-
-    image_barrier_.srcQueueFamilyIndex = src;
-    image_barrier_.dstQueueFamilyIndex = dst;
-    buffer_barrier_.srcQueueFamilyIndex = src;
-    buffer_barrier_.dstQueueFamilyIndex = dst;
-
-    QueueFamilyObjs *qf = GetQueueFamilyInfo(context_, queue_family_index);
-
-    vkt::CommandBuffer *command_buffer = qf->command_buffer;
-    for (int cb_repeat = 0; cb_repeat < (mod == Modifier::DOUBLE_COMMAND_BUFFER ? 2 : 1); cb_repeat++) {
-        command_buffer->begin();
-        for (int repeat = 0; repeat < (mod == Modifier::DOUBLE_RECORD ? 2 : 1); repeat++) {
-            vk::CmdPipelineBarrier(command_buffer->handle(), VK_PIPELINE_STAGE_TRANSFER_BIT, VK_PIPELINE_STAGE_TRANSFER_BIT,
-                                   VK_DEPENDENCY_BY_REGION_BIT, 0, nullptr, 1, &buffer_barrier_, 1, &image_barrier_);
-        }
-        command_buffer->end();
-        command_buffer = qf->command_buffer2;  // Second pass (if any) goes to the secondary command_buffer.
-    }
-
-    if (queue_family_index != kInvalidQueueFamily) {
-        if (mod == Modifier::DOUBLE_COMMAND_BUFFER) {
-            // the Fence resolves to VK_NULL_HANLE... i.e. no fence
-            qf->queue->submit({{qf->command_buffer, qf->command_buffer2}}, vkt::Fence(), positive);
-        } else {
-            qf->command_buffer->QueueCommandBuffer(positive);  // Check for success on positive tests only
-        }
-    }
-
-    if (!positive) {
-        monitor.VerifyFound();
-    }
-    context_->Reset();
-}
-
-void Barrier2QueueFamilyTestHelper::operator()(const std::string &img_err, const std::string &buf_err, uint32_t src, uint32_t dst,
-                                               uint32_t queue_family_index, Modifier mod) {
-    auto &monitor = context_->layer_test->Monitor();
-    bool positive = true;
-    if (img_err.length()) {
-        monitor.SetDesiredFailureMsg(kErrorBit | kWarningBit, img_err);
-        positive = false;
-    }
-    if (buf_err.length()) {
-        monitor.SetDesiredFailureMsg(kErrorBit | kWarningBit, buf_err);
-        positive = false;
-    }
-
-    image_barrier_.srcQueueFamilyIndex = src;
-    image_barrier_.dstQueueFamilyIndex = dst;
-    buffer_barrier_.srcQueueFamilyIndex = src;
-    buffer_barrier_.dstQueueFamilyIndex = dst;
-
-    VkDependencyInfoKHR dep_info = vku::InitStructHelper();
-    dep_info.bufferMemoryBarrierCount = 1;
-    dep_info.pBufferMemoryBarriers = &buffer_barrier_;
-    dep_info.imageMemoryBarrierCount = 1;
-    dep_info.pImageMemoryBarriers = &image_barrier_;
-
-    QueueFamilyObjs *qf = GetQueueFamilyInfo(context_, queue_family_index);
-
-    vkt::CommandBuffer *command_buffer = qf->command_buffer;
-    for (int cb_repeat = 0; cb_repeat < (mod == Modifier::DOUBLE_COMMAND_BUFFER ? 2 : 1); cb_repeat++) {
-        command_buffer->begin();
-        for (int repeat = 0; repeat < (mod == Modifier::DOUBLE_RECORD ? 2 : 1); repeat++) {
-            vk::CmdPipelineBarrier2KHR(command_buffer->handle(), &dep_info);
-        }
-        command_buffer->end();
-        command_buffer = qf->command_buffer2;  // Second pass (if any) goes to the secondary command_buffer.
-    }
-
-    if (queue_family_index != kInvalidQueueFamily) {
-        if (mod == Modifier::DOUBLE_COMMAND_BUFFER) {
-            // the Fence resolves to VK_NULL_HANLE... i.e. no fence
-            qf->queue->submit({{qf->command_buffer, qf->command_buffer2}}, vkt::Fence(), positive);
-        } else {
-            qf->command_buffer->QueueCommandBuffer(positive);  // Check for success on positive tests only
-        }
-    }
-
-    if (!positive) {
-        monitor.VerifyFound();
-    }
-    context_->Reset();
-}
-
-void VkSyncValTest::InitSyncValFramework(bool enable_queue_submit_validation) {
-    // Enable synchronization validation
-
-    // Optional feature definition, add if requested (but they can't be defined at the conditional scope)
-    const char *kEnableQueuSubmitSyncValidation[] = {"VALIDATION_CHECK_ENABLE_SYNCHRONIZATION_VALIDATION_QUEUE_SUBMIT"};
-    const VkLayerSettingEXT settings[] = {
-        {OBJECT_LAYER_NAME, "enables", VK_LAYER_SETTING_TYPE_STRING_EXT, 1, kEnableQueuSubmitSyncValidation}};
-    const VkLayerSettingsCreateInfoEXT qs_settings{VK_STRUCTURE_TYPE_LAYER_SETTINGS_CREATE_INFO_EXT, nullptr,
-                                                   static_cast<uint32_t>(std::size(settings)), settings};
-
-    if (enable_queue_submit_validation) {
-        features_.pNext = &qs_settings;
-    }
-    InitFramework(&features_);
 }
 
 void print_android(const char *c) {
