+++ tests/unit/best_practices_positive.cpp
@@ -13,6 +13,7 @@
 
 #include "../framework/layer_validation_tests.h"
 #include "../framework/pipeline_helper.h"
+#include "../framework/descriptor_helper.h"
 
 class VkPositiveBestPracticesLayerTest : public VkBestPracticesLayerTest {};
