+++ tests/unit/object_lifetime.cpp
@@ -16,6 +16,7 @@
 #include "generated/enum_flag_bits.h"
 #include "../framework/layer_validation_tests.h"
 #include "../framework/pipeline_helper.h"
+#include "../framework/descriptor_helper.h"
 
 // Validation of dispatchable handles is not performed until VVL's chassis will be
 // able to do this validation (if ever) instead of crashing (which is also an option).
