+++ tests/unit/pipeline_positive.cpp
@@ -13,6 +13,7 @@
 
 #include "../framework/layer_validation_tests.h"
 #include "../framework/pipeline_helper.h"
+#include "../framework/descriptor_helper.h"
 #include "generated/vk_extension_helper.h"
 
 TEST_F(PositivePipeline, ComplexTypes) {
