+++ tests/unit/push_descriptor.cpp
@@ -16,6 +16,7 @@
 #include "utils/cast_utils.h"
 #include "../framework/layer_validation_tests.h"
 #include "../framework/pipeline_helper.h"
+#include "../framework/descriptor_helper.h"
 
 TEST_F(NegativePushDescriptor, DSBufferInfo) {
     TEST_DESCRIPTION(
