+++ tests/unit/threading_positive.cpp
@@ -13,6 +13,7 @@
  */
 
 #include "../framework/layer_validation_tests.h"
+#include "../framework/descriptor_helper.h"
 
 #if GTEST_IS_THREADSAFE
 TEST_F(PositiveThreading, DisplayObjects) {
