+++ tests/unit/ycbcr.cpp
@@ -15,6 +15,7 @@
 #include "utils/cast_utils.h"
 #include "generated/enum_flag_bits.h"
 #include "../framework/layer_validation_tests.h"
+#include "../framework/descriptor_helper.h"
 #include "utils/vk_layer_utils.h"
 
 TEST_F(NegativeYcbcr, Sampler) {
